// types/types.ts
export interface Period {
  id: string;
  year: number;
  term: number;
  startDate: Date;
  endDate: Date;
  isActive: boolean;
  schoolId: string;
}

export interface GroupedPeriods {
  [year: string]: Period[];
}
export type DepartmentBrief = {
  id: string;
  name: string;
};
export type ClassBrief = {
  id: string;
  title: string;
};
export type SubjectBrief = {
  id: string;
  name: string;
};

export type StreamWithCount = {
  id: string;
  title: string;
  slug: string;
  classId: string;
  _count: {
    students: number;
  };
  createdAt: string;
  updatedAt: string;
};
export type Stream = {
  id: string;
  title: string;
  slug: string;
  classId: string;
  class: Class;
  createdAt: Date;
  updatedAt: Date;
};
export interface MarkSheetStudent {
  id: string;
  name: string;
}
export interface MarkSheetResponse {
  students: MarkSheetStudent[];
  markSheetId: string;
}
export type CreateMarkSheetProps = {
  examId: string;
  title: string;
  classId: string;
  subjectId: string;
  markSheetId: string;
  termId: string;
};
export type UpdateMarkSheetProps = {
  examId: string;
  classId: string;
  subjectId: string;
  markSheetId: string;
  termId: string;
  studentMarks: {
    studentId: string;
    marks: number | null;
    isAbsent: boolean;
    comments: string;
  }[];
};
// ...
export type Exam = {
  id: string;
  title: string;
  examType: "REGULAR" | "MOCK" | "PLACEMENT";
  examCategory: "TERM_START" | "MID_TERM" | "END_TERM";
  termName: number;
  academicYear: string;
  startDate: string;
};